# server/utils/swagger.yaml
openapi: 3.1.0
info:
  title: Chort Tools Api
  version: 1.0.0
  description: API documation for managing Ironhack cohorts and students
servers:
  - url: http://localhost:5005/
paths:
...
  /auth/signup:
    post:
      summary: Creates a new user
      responses:
        "201":
          description: New user created
        "500":
          description: Internal server error
  /auth/login:
    post:
      summary:  Checks user credentials and creates an auth token
      responses:
        "200":
          description: Credentials match, token created
        "401":
          description: Unable to authenticate user
        "500": 
          description: Internal server error
...